changed to a signed 16 bit address field in symbol table

+++ src/include/compiler.h
@@ -7,7 +7,7 @@
 typedef struct label_list {
 	struct label_list* next;
 	char label[MAX];
-	int32_t address;
+	int16_t address;
 	int size;
 	// used only with the dc keyword
 	int value;
